+++ test/processors/android/android_manifest_processor_test.dart
@@ -25,7 +25,7 @@
 
 import 'dart:io';
 
-import 'package:flutter_flavorizr/parser/models/pubspec.dart';
+import 'package:flutter_flavorizr/parser/models/flavorizr.dart';
 import 'package:flutter_flavorizr/parser/parser.dart';
 import 'package:flutter_flavorizr/processors/android/android_manifest_processor.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -33,12 +33,12 @@ import 'package:flutter_test/flutter_test.dart';
 import '../../test_utils.dart';
 
 void main() {
-  Pubspec? pubspec;
+  late Flavorizr flavorizr;
 
   setUp(() {
-    Parser parser = Parser(file: 'test_resources/pubspec.yaml');
+    Parser parser = Parser(pubspecPath: 'test_resources/pubspec.yaml', flavorizrPath: '',);
     try {
-      pubspec = parser.parse();
+      flavorizr = parser.parse();
     } catch (e) {
       fail(e.toString());
     }
@@ -55,7 +55,7 @@ void main() {
         .readAsStringSync();
 
     AndroidManifestProcessor processor =
-        AndroidManifestProcessor(input: content, config: pubspec!.flavorizr);
+        AndroidManifestProcessor(input: content, config: flavorizr);
     String actual = processor.execute();
 
     actual = TestUtils.stripEndOfLines(actual);
@@ -67,7 +67,7 @@ void main() {
   test('Test malformed AndroidManifestProcessor', () {
     AndroidManifestProcessor processor = AndroidManifestProcessor(
       input: '',
-      config: pubspec!.flavorizr,
+      config: flavorizr,
     );
     expect(() => processor.execute(), throwsException);
   });