+++ CHANGELOG.md
@@ -1,22 +1,29 @@
 ## NEXT
+
+* Changed the way flutter_flavorizr reads the instructions: from the flavorizr.yaml as a main source
+  and pubspec.yaml as fallback
 * Android and iOS definitions are no longer required
 * main.dart generated more seamlessly
 
 ## 2.1.6
+
 * Create file path if not existing when copying icon sets
 * Set letters lowercase when created flavors
 * Bumped dependencies
 * Updated copyright version
 
 ## 2.1.5
+
 * Json serializable build runner setup
 * Bumped dependencies
 
 ## 2.1.4
+
 * Added CFBundleDisplayName in Info.plist to show the proper name for iOS apps
 * Bumped dependencies
 
 ## 2.1.3
+
 * Added Android custom properties/config to flavors
 * Relaxed first app entry checks
 * New naming conventions for flutter_lints
@@ -24,15 +31,18 @@
 * Bumped dependencies
 
 ## 2.1.2
+
 * Added resValues directive to declare resValues in build.gradle for each flavor (Android)
 * Added variables directive to declare variables in xcconfig for each flavor (iOS)
 * Bumped dependencies
 
 ## 2.1.1
+
 * Added icon directive to generate icons for each flavor
 * Bumped dependencies
 
 ## 2.1.0
+
 * Added flavor banner in the top-left corner
 * Disabled override of the launchScreen attribute in Info.plist if its processor is not specified
 * Removed blank line generation at the end of the build.gradle file on every run
@@ -40,41 +50,52 @@
 * Bumped dependencies
 
 ## 2.0.0
+
 * Migration to Flutter 2
 
 ## 1.0.11
+
 * Added FirebaseProcessor to plug Firebase configurations
 
 ## 1.0.10
+
 * Added AndroidBuildGradleProcessor idempotency
 * Bumped dependencies
 * Updated LICENCE
 
 ## 1.0.9
+
 * Fixed README
 
 ## 1.0.8
+
 * Added buildConfig param in iOS configuration
 * Added program argument in vscode configuration
 * Added dash in iOS configuration
 
 ## 1.0.7
+
 * Bumped dependencies
 
 ## 1.0.6
+
 * Bumped dependencies
 
 ## 1.0.5
+
 * Bumped dependencies
 
 ## 1.0.4
+
 * Added processors param to the command line
 
 ## 1.0.3
+
 * Added IDE processors (vscode and idea)
 * Updated README
 
 ## 1.0.2
+
 * Fixed zip asset extraction using sync
 * Fixed zip asset creation with recursive param